Implement OAuth2 and deprecate email/password (#118)
`gmusicapi` has deprecated email/password login and encourages use of
OAuth2 for authorization. As long as an email or password is not
specified in the configuration on the command-line, we attempt to
perform an oauth login. `gmusicapi` caches credentials to disk after a
successful authorization, so we use those if they exist, otherwise, we
prompt the user to login and authorize. This means that on a first run,
we open a browser (or just show a URL) to let the user login.